+++ Z/Data/Vector/Sort.hs
@@ -1,14 +1,3 @@
-{-# LANGUAGE BangPatterns #-}
-{-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE UnboxedTuples #-}
-{-# LANGUAGE MagicHash #-}
-{-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE GeneralizedNewtypeDeriving #-}
-{-# LANGUAGE PolyKinds #-}
-{-# LANGUAGE DataKinds #-}
-{-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE CPP #-}
-
 {-|
 Module      : Z.Data.Vector.Sort
 Description : Sorting vectors
@@ -68,6 +57,7 @@ import           Data.Primitive         (sizeOf)
 import           Data.Word
 import           Prelude                hiding (splitAt)
 import           Z.Data.Array
+import           Z.Data.Array.Unaligned
 import           Z.Data.Vector.Base
 import           Z.Data.Vector.Extra
 
@@ -97,6 +87,8 @@ mergeSortBy cmp vec@(Vec _ _ l)
         w <- mergePass w1 w2 mergeTileSize
         return $! fromArr w 0 l)
   where
+    firstPass :: forall s. v a -> Int -> MArr (IArray v) s a -> ST s ()
+    {-# INLINABLE firstPass #-}
     firstPass !v !i !marr
         | i >= l     = return ()
         | otherwise = do
@@ -104,12 +96,16 @@ mergeSortBy cmp vec@(Vec _ _ l)
             insertSortToMArr cmp v' i marr
             firstPass rest (i+mergeTileSize) marr
 
+    mergePass :: forall s. MArr (IArray v) s a -> MArr (IArray v) s a -> Int -> ST s (IArray v a)
+    {-# INLINABLE mergePass #-}
     mergePass !w1 !w2 !blockSiz
         | blockSiz >= l = unsafeFreezeArr w1
         | otherwise     = do
             mergeLoop w1 w2 blockSiz 0
             mergePass w2 w1 (blockSiz*2) -- swap worker array and continue merging
 
+    mergeLoop :: forall s. MArr (IArray v) s a -> MArr (IArray v) s a -> Int -> Int -> ST s ()
+    {-# INLINABLE mergeLoop #-}
     mergeLoop !src !target !blockSiz !i
         | i >= l-blockSiz =                 -- remaining elements less than a block
             if i >= l
@@ -120,6 +116,8 @@ mergeSortBy cmp vec@(Vec _ _ l)
             mergeBlock src target (i+blockSiz) mergeEnd i (i+blockSiz) i
             mergeLoop src target blockSiz mergeEnd
 
+    mergeBlock :: forall s. MArr (IArray v) s a -> MArr (IArray v) s a -> Int -> Int -> Int -> Int -> Int -> ST s ()
+    {-# INLINABLE mergeBlock #-}
     mergeBlock !src !target !leftEnd !rightEnd !i !j !k = do
         lv <- readArr src i
         rv <- readArr src j
@@ -263,7 +261,8 @@ instance Radix Word64 where MULTI_BYTES_INT_RADIX(Word64)
 
 -- | Similar to 'Down' newtype for 'Ord', this newtype can inverse the order of a 'Radix'
 -- instance when used in 'radixSort'.
-newtype RadixDown a = RadixDown a deriving (Show, Eq, Prim)
+newtype RadixDown a = RadixDown a deriving (Show, Eq)
+                                    deriving newtype (Prim, Unaligned)
 
 instance Radix a => Radix (RadixDown a) where
     {-# INLINE bucketSize #-}
@@ -293,14 +292,14 @@ radixSort v@(Vec arr s l)
     | l <= 1 = v
     | otherwise = runST (do
         bucket <- newArrWith buktSiz 0 :: ST s (MutablePrimArray s Int)
-        w1 <- newArr l
+        w1 <- newArr l :: ST s (MArr (IArray v) s a)
         firstCountPass arr bucket s
         accumBucket bucket buktSiz 0 0
         firstMovePass arr s bucket w1
         w <- if passSiz == 1
             then unsafeFreezeArr w1
             else do
-                w2 <- newArr l
+                w2 <- newArr l :: ST s (MArr (IArray v) s a)
                 radixLoop w1 w2 bucket buktSiz 1
         return $! fromArr w 0 l)
   where
@@ -309,6 +308,7 @@ radixSort v@(Vec arr s l)
     !end = s + l
 
     {-# INLINABLE firstCountPass #-}
+    firstCountPass :: forall s. IArray v a -> MutablePrimArray s Int -> Int -> ST s ()
     firstCountPass !arr' !bucket !i
         | i >= end  = return ()
         | otherwise = case indexArr' arr' i of
@@ -319,6 +319,7 @@ radixSort v@(Vec arr s l)
                 firstCountPass arr' bucket (i+1)
 
     {-# INLINABLE accumBucket #-}
+    accumBucket :: forall s. MutablePrimArray s Int -> Int -> Int -> Int -> ST s ()
     accumBucket !bucket !bsiz !i !acc
         | i >= bsiz = return ()
         | otherwise = do
@@ -327,6 +328,7 @@ radixSort v@(Vec arr s l)
             accumBucket bucket bsiz (i+1) (acc+c)
 
     {-# INLINABLE firstMovePass #-}
+    firstMovePass :: forall s. IArray v a -> Int -> MutablePrimArray s Int -> MArr (IArray v) s a -> ST s ()
     firstMovePass !arr' !i !bucket !w
         | i >= end  = return ()
         | otherwise = case indexArr' arr' i of
@@ -338,6 +340,7 @@ radixSort v@(Vec arr s l)
                 firstMovePass arr' (i+1) bucket w
 
     {-# INLINABLE radixLoop #-}
+    radixLoop :: forall s. MArr (IArray v) s a -> MArr (IArray v) s a -> MutablePrimArray s Int -> Int -> Int -> ST s ((IArray v) a)
     radixLoop !w1 !w2 !bucket !bsiz !pass
         | pass >= passSiz-1 = do
             setArr bucket 0 bsiz 0   -- clear the counting bucket
@@ -353,6 +356,7 @@ radixSort v@(Vec arr s l)
             radixLoop w2 w1 bucket bsiz (pass+1)
 
     {-# INLINABLE countPass #-}
+    countPass :: forall s. MArr (IArray v) s a -> MutablePrimArray s Int -> Int -> Int -> ST s ()
     countPass !marr !bucket !pass !i
         | i >= l  = return ()
         | otherwise = do
@@ -363,6 +367,7 @@ radixSort v@(Vec arr s l)
                 countPass marr bucket pass (i+1)
 
     {-# INLINABLE movePass #-}
+    movePass :: forall s. MArr (IArray v) s a -> MutablePrimArray s Int -> Int -> MArr (IArray v) s a -> Int -> ST s ()
     movePass !src !bucket !pass !target !i
         | i >= l  = return ()
         | otherwise = do
@@ -374,6 +379,7 @@ radixSort v@(Vec arr s l)
                 movePass src bucket pass target (i+1)
 
     {-# INLINABLE lastCountPass #-}
+    lastCountPass :: forall s. MArr (IArray v) s a -> MutablePrimArray s Int -> Int -> ST s ()
     lastCountPass !marr !bucket !i
         | i >= l  = return ()
         | otherwise = do
@@ -384,6 +390,7 @@ radixSort v@(Vec arr s l)
                 lastCountPass marr bucket (i+1)
 
     {-# INLINABLE lastMovePass #-}
+    lastMovePass :: forall s. MArr (IArray v) s a -> MutablePrimArray s Int -> MArr (IArray v) s a -> Int -> ST s ()
     lastMovePass !src !bucket !target !i
         | i >= l  = return ()
         | otherwise = do
@@ -449,12 +456,12 @@ radixSortFloat v =  castVector (radixSort (castVector v :: PrimVector RadixFloat
 -- | merge duplicated adjacent element, prefer left element.
 --
 -- Use this function on a sorted vector will have the same effects as 'nub'.
-mergeDupAdjacent :: (Vec v a, Eq a) => v a -> v a
+mergeDupAdjacent :: forall v a. (Vec v a, Eq a) => v a -> v a
 {-# INLINE mergeDupAdjacent #-}
 mergeDupAdjacent = mergeDupAdjacentBy (==) const
 
 -- | Merge duplicated adjacent element, prefer left element.
-mergeDupAdjacentLeft :: Vec v a
+mergeDupAdjacentLeft :: forall v a. Vec v a
                      => (a -> a -> Bool)   -- ^ equality tester, @\ left right -> eq left right@
                      -> v a
                      -> v a
@@ -462,7 +469,7 @@ mergeDupAdjacentLeft eq = mergeDupAdjacentBy eq const
 {-# INLINE mergeDupAdjacentLeft #-}
 
 -- | Merge duplicated adjacent element, prefer right element.
-mergeDupAdjacentRight :: Vec v a
+mergeDupAdjacentRight :: forall v a. Vec v a
                       => (a -> a -> Bool)  -- ^ equality tester, @\ left right -> eq left right@
                       -> v a
                       -> v a
@@ -470,7 +477,7 @@ mergeDupAdjacentRight :: Vec v a
 mergeDupAdjacentRight eq = mergeDupAdjacentBy eq (\ _ x -> x)
 
 -- | Merge duplicated adjacent element, based on a equality tester and a merger function.
-mergeDupAdjacentBy :: Vec v a
+mergeDupAdjacentBy :: forall v a. Vec v a
                    => (a -> a -> Bool)  -- ^ equality tester, @\ left right -> eq left right@
                    -> (a -> a -> a)     -- ^ the merger, @\ left right -> merge left right@
                    -> v a -> v a
@@ -484,6 +491,7 @@ mergeDupAdjacentBy eq merger v@(Vec arr s l)
         go arr marr s 1 x0
   where
     !end = s + l
+    go :: forall s. IArray v a -> MArr (IArray v) s a -> Int -> Int -> a -> ST s Int
     go !arr' !marr !i !j !x
         | i >= end  = return j
         | otherwise = do